+++ integration/server.js
@@ -1,18 +1,30 @@
 const express = require('express');
 const ParseServer = require('parse-server').ParseServer;
 const app = express();
+const CustomAuth = require('./test/CustomAuth');
 
-// Specify the connection string for your mongodb database
-// and the location to your Parse cloud code
 const api = new ParseServer({
   databaseURI: 'mongodb://localhost:27017/integration',
   appId: 'integration',
   masterKey: 'notsosecret',
-  serverURL: 'http://localhost:1337/parse', // Don't forget to change to https if needed
-  cloud: `${__dirname}/cloud/main.js`
+  serverURL: 'http://localhost:1337/parse',
+  cloud: `${__dirname}/cloud/main.js`,
+  liveQuery: {
+    classNames: ['TestObject', 'DiffObject'],
+  },
+  startLiveQueryServer: true,
+  auth: {
+    myAuth: {
+      module: CustomAuth,
+      option1: 'hello',
+      option2: 'world',
+    },
+    facebook: {
+      appIds: "test"
+    }
+  }
 });
 
-// Serve the Parse API on the /parse URL prefix
 app.use('/parse', api);
 
 const TestUtils = require('parse-server').TestUtils;