+++ integration/test/ParseUserTest.js
@@ -17,6 +17,33 @@ class CustomUser extends Parse.User {
 }
 Parse.Object.registerSubclass('CustomUser', CustomUser);
 
+const provider = {
+  authenticate: () => Promise.resolve(),
+  restoreAuthentication: () => true,
+  getAuthType: () => 'anonymous',
+  getAuthData() {
+    return {
+      authData: {
+        id: '1234',
+      },
+    };
+  },
+};
+Parse.User._registerAuthenticationProvider(provider);
+
+const authResponse = {
+  userID: 'test',
+  accessToken: 'test',
+  expiresIn: 'test', // Should be unix timestamp
+};
+global.FB = {
+  init: () => {},
+  login: (cb) => {
+    cb({ authResponse });
+  },
+  getAuthResponse: () => authResponse,
+};
+
 describe('Parse User', () => {
   beforeAll(() => {
     Parse.initialize('integration', null, 'notsosecret');
@@ -561,5 +588,136 @@ describe('Parse User', () => {
     user = await CustomUser.logIn('username', 'password');
     expect(user instanceof CustomUser).toBe(true);
     expect(user.doSomething()).toBe(5);
+  });
+
+  it('can link without master key', async () => {
+    Parse.User.enableUnsafeCurrentUser();
+
+    const user = new Parse.User();
+    user.setUsername('Alice');
+    user.setPassword('sekrit');
+    await user.signUp();
+    await user._linkWith(provider.getAuthType(), provider.getAuthData());
+    expect(user._isLinked(provider)).toBe(true);
+    await user._unlinkFrom(provider);
+    expect(user._isLinked(provider)).toBe(false);
+  });
+
+  it('can link with master key', async () => {
+    Parse.User.disableUnsafeCurrentUser();
+
+    const user = new Parse.User();
+    user.setUsername('Alice');
+    user.setPassword('sekrit');
+    await user.save(null, { useMasterKey: true });
+    await user._linkWith(provider.getAuthType(), provider.getAuthData(), { useMasterKey: true });
+    expect(user._isLinked(provider)).toBe(true);
+    await user._unlinkFrom(provider, { useMasterKey: true });
+    expect(user._isLinked(provider)).toBe(false);
+  });
+
+  it('can link with session token', async () => {
+    Parse.User.disableUnsafeCurrentUser();
+
+    const user = new Parse.User();
+    user.setUsername('Alice');
+    user.setPassword('sekrit');
+    await user.signUp();
+    expect(user.isCurrent()).toBe(false);
+
+    const sessionToken = user.getSessionToken();
+    await user._linkWith(provider.getAuthType(), provider.getAuthData(), { sessionToken });
+    expect(user._isLinked(provider)).toBe(true);
+    await user._unlinkFrom(provider, { sessionToken });
+    expect(user._isLinked(provider)).toBe(false);
+  });
+
+  it('linked account can login with authData', async () => {
+    const user = new Parse.User();
+    user.setUsername('Alice');
+    user.setPassword('sekrit');
+    await user.save(null, { useMasterKey: true });
+    await user._linkWith(provider.getAuthType(), provider.getAuthData(), { useMasterKey: true });
+    expect(user._isLinked(provider)).toBe(true);
+    expect(user.authenticated()).toBeFalsy();
+    Parse.User.enableUnsafeCurrentUser();
+    const loggedIn = await Parse.User.logInWith(provider.getAuthType(), provider.getAuthData());
+    expect(loggedIn.authenticated()).toBeTruthy();
+  });
+
+  it('linking un-authenticated user without master key will throw', async (done) => {
+    const user = new Parse.User();
+    user.setUsername('Alice');
+    user.setPassword('sekrit');
+    await user.save(null, { useMasterKey: true });
+    user._linkWith(provider.getAuthType(), provider.getAuthData())
+      .then(() => done.fail('should fail'))
+      .catch(e => expect(e.message).toBe(`Cannot modify user ${user.id}.`))
+      .then(done);
+  });
+
+  it('can link with custom auth', async () => {
+    Parse.User.enableUnsafeCurrentUser();
+    const provider = {
+      authenticate: () => Promise.resolve(),
+      restoreAuthentication() {
+        return true;
+      },
+
+      getAuthType() {
+        return 'myAuth';
+      },
+
+      getAuthData() {
+        return {
+          authData: {
+            id: 1234,
+          },
+        };
+      },
+    };
+    Parse.User._registerAuthenticationProvider(provider);
+    const user = new Parse.User();
+    user.setUsername('Alice');
+    user.setPassword('sekrit');
+    await user.signUp();
+    await user._linkWith(provider.getAuthType(), provider.getAuthData());
+    expect(user._isLinked(provider)).toBe(true);
+    await user._unlinkFrom(provider);
+    expect(user._isLinked(provider)).toBe(false);
+  });
+
+  it('can login with facebook', async () => {
+    Parse.User.enableUnsafeCurrentUser();
+    Parse.FacebookUtils.init();
+    const user = await Parse.FacebookUtils.logIn();
+    expect(Parse.FacebookUtils.isLinked(user)).toBe(true);
+  });
+
+  it('can link user with facebook', async () => {
+    Parse.User.enableUnsafeCurrentUser();
+    Parse.FacebookUtils.init();
+    const user = new Parse.User();
+    user.setUsername('Alice');
+    user.setPassword('sekrit');
+    await user.signUp();
+    await Parse.FacebookUtils.link(user);
+    expect(Parse.FacebookUtils.isLinked(user)).toBe(true);
+    await Parse.FacebookUtils.unlink(user);
+    expect(Parse.FacebookUtils.isLinked(user)).toBe(false);
+  });
+
+  it('can link anonymous user with facebook', async () => {
+    Parse.User.enableUnsafeCurrentUser();
+    Parse.FacebookUtils.init();
+    const user = await Parse.AnonymousUtils.logIn();
+    await Parse.FacebookUtils.link(user);
+
+    expect(Parse.FacebookUtils.isLinked(user)).toBe(true);
+    expect(Parse.AnonymousUtils.isLinked(user)).toBe(true);
+    await Parse.FacebookUtils.unlink(user);
+
+    expect(Parse.FacebookUtils.isLinked(user)).toBe(false);
+    expect(Parse.AnonymousUtils.isLinked(user)).toBe(true);
   });
 });